Exception management for PatientController

+++ src/main/java/com/healthcare/main/controller/PatientController.java
@@ -1,8 +1,12 @@
 package com.healthcare.main.controller;
 
+import com.healthcare.main.exception.BadRequestException;
+import com.healthcare.main.exception.NotFoundException;
+import com.healthcare.main.model.Doctor;
 import com.healthcare.main.model.Patient;
 import com.healthcare.main.service.PatientService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,8 +24,13 @@ public class PatientController
     }
 
     @GetMapping(value="/{id}")
-    public Patient getPatient(@PathVariable Long id)
+    public Patient getPatient(@PathVariable("id") Long id) throws NotFoundException
     {
+        Patient patient = patientService.getPatient(id);
+        if(patient == null)
+        {
+            throw new NotFoundException(String.format("Patient with id=%s was not found.", id));
+        }
         return patientService.getPatient(id);
     }
 
@@ -32,6 +41,7 @@ public class PatientController
     }
 
     @PostMapping
+    @ResponseStatus(value = HttpStatus.CREATED)
     public Patient savePatient(@RequestBody Patient patient)
     {
         patientService.savePatient(patient);
@@ -39,28 +49,38 @@ public class PatientController
     }
 
     @PutMapping(value="/{id}")
-    public Patient updatePatient(@RequestBody Patient patient)
+    public Patient updatePatient(@PathVariable("id") Long id, @RequestBody Patient patient) throws BadRequestException, NotFoundException
     {
-        return patientService.updatePatient(patient);
+        if(!id.equals(patient.getPatientID()))
+        {
+            throw new BadRequestException("The id is not the same with id from object");
+        }
+
+        Patient patientDb = patientService.getPatient(id);
+
+        if(patientDb == null)
+        {
+            throw new NotFoundException(String.format("Patient with id=%s was not found.", id));
+        }
+
+        return patientService.updatePatient(patientDb);
     }
 
-//    @PutMapping
-//    public ResponseEntity<Object> updatePatients()
-//    {
-//        // doctorService.updatePatients(patientList);
-//        return new ResponseEntity<>(null, HttpStatus.NOT_IMPLEMENTED);
-//    }
-
     @DeleteMapping(value="/{id}")
-    public Boolean deletePatient(@PathVariable Long id)
+    @ResponseStatus(value = HttpStatus.NO_CONTENT)
+    public void deletePatient(@PathVariable("id") Long id) throws NotFoundException
     {
         Patient patient = patientService.getPatient(id);
-        return patientService.deletePatient(patient);
+        if(patient == null){
+            throw new NotFoundException(String.format("Patient with id=%s was not found.", id));
+        }
+        patientService.deletePatient(patient);
     }
 
     @DeleteMapping
-    public Boolean deleteAllPatients()
+    @ResponseStatus(value = HttpStatus.NO_CONTENT)
+    public void deleteAllPatients()
     {
-        return patientService.deleteAllPatients();
+        patientService.deleteAllPatients();
     }
 }
